+++ packages/router/README.md
@@ -74,6 +74,9 @@ const router = new Router({
 
 	// run at the start of each request, return state to use in middleware
 	start: (c) => ({ foo: "bar" }),
+
+	// base `Page` to inject elements into, this is the default
+	page: '<!doctype html><html><head><meta charset="UTF-8"><meta name="viewport" content="width=device-width, initial-scale=1.0"></head><body></body></html>',
 });
 ```
