+++ packages/router/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./trie/index.js";
 export * from "./fetch/index.js";
+export type * from "./types.js";
